requirements.txt
Django==1.4.1
South==0.7.6
argparse==1.2.1
distribute==0.6.24
django-extensions==0.9
ipython==0.13
pyserial==2.6
wsgiref==0.1.2
